# Spreading score optimization: Store ignored instead of qualifying nodes